+++ projects/_07_workforce_characteristics/_01_starters_leavers_vacancies/fargate/_01_merge.py
@@ -76,9 +76,6 @@ def main(
     )
 
     # TODO: Placeholder only
-    # mUtils.convert_ascwds_job_role_columns_to_rows()
-
-    # TODO: Placeholder only
     # mUtils.join_datasets()
 
     # TODO: Placeholder only