+++ projects/_07_workforce_characteristics/_01_starters_leavers_vacancies/tests/fargate/test_01_merge.py
@@ -15,14 +15,12 @@ class MainTests(unittest.TestCase):
     @patch(f"{PATCH_PATH}.utils.sink_to_parquet")
     @patch(f"{PATCH_PATH}.mUtils.apply_employment_status_magic_numbers")
     @patch(f"{PATCH_PATH}.mUtils.join_datasets")
-    @patch(f"{PATCH_PATH}.mUtils.convert_ascwds_job_role_columns_to_rows")
     @patch(f"{PATCH_PATH}.expr.is_slv_job_role_column")
     @patch(f"{PATCH_PATH}.utils.scan_parquet")
     def test_main_runs(
         self,
         scan_parquet_mock: Mock,
         is_slv_job_role_column_mock: Mock,
-        convert_ascwds_job_role_columns_to_rows_mock: Mock,
         join_datasets_mock: Mock,
         apply_employment_status_magic_numbers_mock: Mock,
         sink_to_parquet_mock: Mock,
@@ -48,7 +46,6 @@ class MainTests(unittest.TestCase):
 
         # TODO: Uncomment these assertions when the placeholder functions are implemented
         is_slv_job_role_column_mock.assert_called_once()
-        # convert_ascwds_job_role_columns_to_rows_mock.assert_called_once()
         # join_datasets_mock.assert_called_once()
         # apply_employment_status_magic_numbers_mock.assert_called_once()